perf(rocm): apply BF16 kernel optimizations to CausalWan I2V model
Mirror the T2V AMD optimizations in CausalWanTransformerBlock:
- Remove hidden_states.float() pre-cast in norm1 to enable AITER BF16
  layernorm path (~5x faster than FP32 F.layer_norm on MI300X)
- Cast FP32 scale_msa/shift_msa to BF16 before large [B,S,C] multiply
  to avoid FP32 intermediate tensor (~2.4x faster modulation op)
- Replace two separate _apply_rotary_emb Triton calls with fused AITER
  rope_cached_2c_fwd for Q+K (~3.5x faster on AMD ROCm)
- Fix torch.zeroes typo → torch.zeros in null_shift/null_scale
- Fix VSA block null_shift/null_scale dtype: add dtype=hidden_states.dtype
  to torch.zeros to avoid spurious FP32 scale/shift intermediates

These changes mirror commits e831e66, 80d78d4, 2037798, and 25e117a
which applied the same optimizations to WanTransformerBlock (T2V).

+++ python/sglang/multimodal_gen/runtime/models/dits/wanvideo.py
@@ -585,7 +585,9 @@ class WanTransformerBlock_VSA(nn.Module):
         attn_output, _ = self.to_out(attn_output)
         attn_output = attn_output.squeeze(1)
 
-        null_shift = null_scale = torch.zeros((1,), device=hidden_states.device)
+        null_shift = null_scale = torch.zeros(
+            (1,), device=hidden_states.device, dtype=hidden_states.dtype
+        )
         norm_hidden_states, hidden_states = self.self_attn_residual_norm(
             hidden_states, attn_output, gate_msa, null_shift, null_scale
         )
